맵에 마커 만들기

+++ src/main/java/com/sample/camping/vo/CampingSites.java
@@ -7,8 +7,8 @@ public class CampingSites {
 	private int no;
 	private	String name;
 	private String sort;
-	private int latitude;
-	private int longitude;
+	private Double latitude;
+	private Double longitude;
 	private String address;
 	private String tel;
 	private int sites;
@@ -49,21 +49,25 @@ public class CampingSites {
 		this.sort = sort;
 	}
 
-	public int getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(int latitude) {
+
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public int getLongitude() {
+
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(int longitude) {
+
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
+
 
 	public String getAddress() {
 		return address;
